+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- moved mex.common.extract.parse_csv to mex.extractors.utils to rid mex-common of pandas
 - add final outbound checks for synopse and voxco
 - tests: integration tests with mex-backend in CI. Skipped locally to avoid manual
   backend setup, will be automated in MX-1523.